*** empty log message ***

+++ coders/app1.c
@@ -76,8 +76,7 @@ static unsigned int
 %
 %  The format of the IsAPP1 method is:
 %
-%      unsigned int IsAPP1(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int IsAPP1(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -90,8 +89,7 @@ static unsigned int
 %
 %
 */
-static unsigned int IsAPP1(const unsigned char *magick,
-  const unsigned int length)
+static unsigned int IsAPP1(const unsigned char *magick,const size_t length)
 {
   if (length < 4)
     return(False);
